+++ lib/commands/version_commands.sh
@@ -8,31 +8,36 @@ version_command() {
   fi
 
   shift 2
-  local versions=$@
+  local versions=("$@")
 
   local file
   if [ $cmd = "global" ]; then
     file=${ASDF_DEFAULT_TOOL_VERSIONS_FILENAME:-$HOME/.tool-versions}
   else
-    file=$(pwd)/.tool-versions
+    file="$(pwd)/.tool-versions"
   fi
 
-  check_if_plugin_exists $plugin
-  for version in $versions; do
-    check_if_version_exists $plugin $version
+  check_if_plugin_exists "$plugin"
+
+  local version
+  for version in "${versions[@]}"; do
+    check_if_version_exists "$plugin" "$version"
   done
 
-  if [ -f "$file" ] && grep $plugin "$file" > /dev/null; then
-    sed -i -e "s/$plugin .*/$plugin $versions/" "$file"
+  if [ -f "$file" ] && grep "^$plugin " "$file" > /dev/null; then
+    sed -i.bak -e "s/^$plugin .*$/$plugin ${versions[*]}/" "$file"
+    rm "$file".bak
   else
-    echo "$plugin $versions" >> "$file"
+    echo "$plugin ${versions[*]}" >> "$file"
   fi
 }
 
 local_command() {
+  # shellcheck disable=2068
   version_command "local" $@
 }
 
 global_command() {
+  # shellcheck disable=2068
   version_command "global" $@
 }