+++ release.sh
@@ -6,6 +6,7 @@ set -euo pipefail
 #ORIGINAL_IFS=$IFS
 IFS=$'\t\n' # Stricter IFS settings
 
+# shellcheck disable=SC2006
 usage() {
     cat <<EOF
 
@@ -17,6 +18,23 @@ versions of asdf. Plugin developers and asdf users do not need this script.
 This script updates the hardcoded versions in the source code and README and
 then commits them on the current branch. It then tags that commit with the
 specified version.
+
+If you run this script in error, or with the wrong version, you can undo the
+changes by finding the original state in the list of actions listed in the
+reflog:
+
+    git reflog
+
+Then revert to the original state by running `git checkout` with the reference
+previous to the release tagging changes:
+
+    git checkout HEAD@{21}
+
+Then checkout the original branch again:
+
+    git checkout master
+
+You are back to the original state!
 
 EOF
 }
@@ -49,6 +67,12 @@ if git tag | grep "$new_tag_name" > /dev/null; then
     exit 1
 fi
 
+# Make sure the changelog already contains details on the new version
+if ! grep "$new_version$" CHANGELOG.md; then
+    echo >&2 "ERROR: changelog entry for this version is missing"
+    exit 1
+fi
+
 echo "INFO: Checking that all changes are commited and pushed"
 git pull
 
@@ -65,16 +89,17 @@ if ! git diff-index --cached --exit-code -r --ignore-submodules HEAD -- >&2; the
 fi
 
 # Update version in README
-sed -i.bak "s|^\(git clone.*--branch \).*$|\1$new_tag_name|" README.md
+sed -i.bak "s|^\\(git clone.*--branch \\).*$|\\1$new_tag_name|" README.md
+rm README.md.bak
 
-# Update version in utils.sh
-# TODO: Hardcode version until.sh since not all asdf will include the Git repo
+# Update version in the VERSION file
+echo "$new_tag_name" > VERSION
 
 echo "INFO: Committing and tagging new version"
 
 # Commit the changed files before tagging the new release
 git add README.md
-git add lib/utils.sh
+git add VERSION
 git commit -m "Update version to $new_version"
 
 git tag -a "$new_tag_name" -m "Version ${new_version}"